gd32v: Add basic support for GD's RISC-V based MCU
GigaDevices now ships a GD32VF103 series MCUs. Most peripherals are similar
to those of GD32F103 (no "V" there) or, for that matter, STM32F103, but a
Bumblebee RISC-V core instead of Cortex-M3.

This commit adds support for parts that are not common with Cortex-M3
based MCUs: the early startup, interrupt handling and toolchain support.

--
Changes since v1:
- Drop the linker script
- Don't select the riscv- toolchain in top-level Makefile
- Tidy up doxygen documentation a bit