make: use vpath only for *.c files
Do not reuse '*.o' files build for one family in another. When GD32V chips are
added, the mixing the Arm and RISC-V code is going to break the build.

But, perhaps even worse, this could cause trouble even when all the files we
build are Arm: an object file build with wrong #defines or compiler flags can
sneak in a library for another chip that happens to share the same code without
notice.

--
Changes since v1:
- Added this patch